+++ pyframe/regions.py
@@ -125,8 +125,6 @@ class Region(object):
         # TODO define standard region specs
         self._name = name
         self._fragments = fragments
-        self._is_inner_region = False
-        self._is_outer_region = False
         self._use_multipoles = False
         self._use_polarizabilities = False
         self._use_fragment_densities = False
@@ -134,33 +132,18 @@ class Region(object):
         self._use_standard_potentials = False
         self._use_lennard_jones = False
         self._use_mfcc = False
+        # general options
+        self._program = 'dalton'
+        self._method = 'DFT'
+        self._xcfun = 'B3LYP'
+        self._basis = 'loprop-6-31+G*'
         # multipoles options
         self._multipole_order = 2
-        self._multipole_program = 'dalton'
-        self._multipole_model = 'loprop'
-        self._multipole_method = 'DFT'
-        self._multipole_xcfun = 'B3LYP'
-        self._multipole_basis = 'loprop-6-31+G*'
         # polarizability options
         self._polarizability_order = (1, 1)
-        self._polarizability_program = 'dalton'
-        self._polarizability_model = 'loprop'
-        self._polarizability_method = 'DFT'
-        self._polarizability_xcfun = 'B3LYP'
-        self._polarizability_basis = 'loprop-6-31+G*'
         self._isotropic_polarizabilities = False
         # fragment_densities options
-        self._fragment_density_program = 'dalton'
-        self._fragment_density_model = 'pde'
-        self._fragment_density_method = 'DFT'
-        self._fragment_density_xcfun = 'B3LYP'
-        self._fragment_density_basis = '6-31+G*'
         # exchange_repulsion options
-        self._exchange_repulsion_program = 'dalton'
-        self._exchange_repulsion_model = 'pde'
-        self._exchange_repulsion_method = 'DFT'
-        self._exchange_repulsion_xcfun = 'B3LYP'
-        self._exchange_repulsion_basis = '6-31+G*'
         self._exchange_repulsion_factor = 0.8
         # standard potentials options
         self._standard_potential_model = 'sep'
@@ -175,19 +158,6 @@ class Region(object):
             else:
                 # TODO: replace exit with exception
                 exit('ERROR: unknown region property "{0}"'.format(key))
-        if self.is_inner_region:
-            if self.is_outer_region:
-                # TODO replace with exception
-                exit('ERROR: inner region and outer region are mutually exclusive')
-            self.use_fragment_densities = True
-            self.use_exchange_repulsion = True
-            self.use_polarizabilities = True
-        if self.is_outer_region:
-            if self.is_inner_region:
-                # TODO replace with exception
-                exit('ERROR: inner region and outer region are mutually exclusive')
-            self.use_multipoles = True
-            self.use_polarizabilities = True
 
     @property
     def name(self):
@@ -205,24 +175,6 @@ class Region(object):
     @fragments.setter
     def fragments(self, fragments):
         self._fragments = fragments
-
-    @property
-    def is_inner_region(self):
-        return self._is_inner_region
-
-    @is_inner_region.setter
-    def is_inner_region(self, is_inner_region):
-        assert isinstance(is_inner_region, bool)
-        self._is_inner_region = is_inner_region
-
-    @property
-    def is_outer_region(self):
-        return self._is_outer_region
-
-    @is_outer_region.setter
-    def is_outer_region(self, is_outer_region):
-        assert isinstance(is_outer_region, bool)
-        self._is_outer_region = is_outer_region
 
     @property
     def use_multipoles(self):
@@ -243,49 +195,40 @@ class Region(object):
         self._multipole_order = order
 
     @property
-    def multipole_program(self):
-        return self._multipole_program
+    def program(self):
+        return self._program
 
-    @multipole_program.setter
-    def multipole_program(self, multipole_program):
-        assert isinstance(multipole_program, str)
-        self._multipole_program = multipole_program.lower()
-
-    @property
-    def multipole_model(self):
-        return self._multipole_model
-
-    @multipole_model.setter
-    def multipole_model(self, multipole_model):
-        assert isinstance(multipole_model, str)
-        self._multipole_model = multipole_model.lower()
+    @program.setter
+    def program(self, program):
+        assert isinstance(program, str)
+        self._program = program.lower()
 
     @property
-    def multipole_method(self):
-        return self._multipole_method
+    def method(self):
+        return self._method
 
-    @multipole_method.setter
-    def multipole_method(self, method):
+    @method.setter
+    def method(self, method):
         assert isinstance(method, str)
-        self._multipole_method = method
+        self._method = method
 
     @property
-    def multipole_xcfun(self):
-        return self._multipole_xcfun
+    def xcfun(self):
+        return self._xcfun
 
-    @multipole_xcfun.setter
-    def multipole_xcfun(self, xcfun):
+    @xcfun.setter
+    def xcfun(self, xcfun):
         assert isinstance(xcfun, str)
-        self._multipole_xcfun = xcfun
+        self._xcfun = xcfun
 
     @property
-    def multipole_basis(self):
-        return self._multipole_basis
+    def basis(self):
+        return self._basis
 
-    @multipole_basis.setter
-    def multipole_basis(self, basis):
+    @basis.setter
+    def basis(self, basis):
         assert isinstance(basis, str)
-        self._multipole_basis = basis
+        self._basis = basis
 
     @property
     def use_polarizabilities(self):
@@ -306,51 +249,6 @@ class Region(object):
         self._polarizability_order = order
 
     @property
-    def polarizability_program(self):
-        return self._polarizability_program
-
-    @polarizability_program.setter
-    def polarizability_program(self, polarizability_program):
-        assert isinstance(polarizability_program, str)
-        self._polarizability_program = polarizability_program.lower()
-
-    @property
-    def polarizability_model(self):
-        return self._polarizability_model
-
-    @polarizability_model.setter
-    def polarizability_model(self, polarizability_model):
-        assert isinstance(polarizability_model, str)
-        self._polarizability_model = polarizability_model.lower()
-
-    @property
-    def polarizability_method(self):
-        return self._polarizability_method
-
-    @polarizability_method.setter
-    def polarizability_method(self, method):
-        assert isinstance(method, str)
-        self._polarizability_method = method
-
-    @property
-    def polarizability_xcfun(self):
-        return self._polarizability_xcfun
-
-    @polarizability_xcfun.setter
-    def polarizability_xcfun(self, xcfun):
-        assert isinstance(xcfun, str)
-        self._polarizability_xcfun = xcfun
-
-    @property
-    def polarizability_basis(self):
-        return self._polarizability_basis
-
-    @polarizability_basis.setter
-    def polarizability_basis(self, basis):
-        assert isinstance(basis, str)
-        self._polarizability_basis = basis
-
-    @property
     def isotropic_polarizabilities(self):
         return self._isotropic_polarizabilities
 
@@ -369,51 +267,6 @@ class Region(object):
         self._use_fragment_densities = use_fragments_densities
 
     @property
-    def fragment_density_program(self):
-        return self._fragment_density_program
-
-    @fragment_density_program.setter
-    def fragment_density_program(self, fragment_density_program):
-        assert isinstance(fragment_density_program, str)
-        self._fragment_density_program = fragment_density_program.lower()
-
-    @property
-    def fragment_density_model(self):
-        return self._fragment_density_model
-
-    @fragment_density_model.setter
-    def fragment_density_model(self, fragment_density_model):
-        assert isinstance(fragment_density_model, str)
-        self._fragment_density_model = fragment_density_model.lower()
-
-    @property
-    def fragment_density_method(self):
-        return self._fragment_density_method
-
-    @fragment_density_method.setter
-    def fragment_density_method(self, method):
-        assert isinstance(method, str)
-        self._fragment_density_method = method
-
-    @property
-    def fragment_density_xcfun(self):
-        return self._fragment_density_xcfun
-
-    @fragment_density_xcfun.setter
-    def fragment_density_xcfun(self, xcfun):
-        assert isinstance(xcfun, str)
-        self._fragment_density_xcfun = xcfun
-
-    @property
-    def fragment_density_basis(self):
-        return self._fragment_density_basis
-
-    @fragment_density_basis.setter
-    def fragment_density_basis(self, basis):
-        assert isinstance(basis, str)
-        self._fragment_density_basis = basis
-
-    @property
     def use_exchange_repulsion(self):
         return self._use_exchange_repulsion
 
@@ -421,51 +274,6 @@ class Region(object):
     def use_exchange_repulsion(self, use_exchange_repulsion):
         assert isinstance(use_exchange_repulsion, bool)
         self._use_exchange_repulsion = use_exchange_repulsion
-
-    @property
-    def exchange_repulsion_program(self):
-        return self._exchange_repulsion_program
-
-    @exchange_repulsion_program.setter
-    def exchange_repulsion_program(self, exchange_repulsion_program):
-        assert isinstance(exchange_repulsion_program, str)
-        self._exchange_repulsion_program = exchange_repulsion_program.lower()
-
-    @property
-    def exchange_repulsion_model(self):
-        return self._exchange_repulsion_model.lower()
-
-    @exchange_repulsion_model.setter
-    def exchange_repulsion_model(self, exchange_repulsion_model):
-        assert isinstance(exchange_repulsion_model, str)
-        self._exchange_repulsion_model = exchange_repulsion_model
-
-    @property
-    def exchange_repulsion_method(self):
-        return self._exchange_repulsion_method
-
-    @exchange_repulsion_method.setter
-    def exchange_repulsion_method(self, method):
-        assert isinstance(method, str)
-        self._exchange_repulsion_method = method
-
-    @property
-    def exchange_repulsion_xcfun(self):
-        return self._exchange_repulsion_xcfun
-
-    @exchange_repulsion_xcfun.setter
-    def exchange_repulsion_xcfun(self, xcfun):
-        assert isinstance(xcfun, str)
-        self._exchange_repulsion_xcfun = xcfun
-
-    @property
-    def exchange_repulsion_basis(self):
-        return self._exchange_repulsion_basis
-
-    @exchange_repulsion_basis.setter
-    def exchange_repulsion_basis(self, basis):
-        assert isinstance(basis, str)
-        self._exchange_repulsion_basis = basis
 
     @property
     def exchange_repulsion_factor(self):